Work in progress on making symmetric permutations work properly.
I need to make sure that the lattice being generated has the correct rotation
symmetries.  This is being done by making sure that the origin of the lattice
is *not* at the seed, but at the center of a maximally symmetric face.  This
origin is a parameter of the tiling, and I have computed it for some (but not
yet all) of the tilings.  Those for which I have added it seem to work correctly
with the 45 degree rotation symmetries, though the square shape of the growth
screws up the 30 degree rotation symmetries.